LayoutTests/storage/indexeddb/mozilla/key-requirements-inline-and-passed-expected.txt

Test IndexedDB's behavior adding inline and passed key simultaneously

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'webkitIndexedDB' in window is true
PASS webkitIndexedDB == null is false
webkitIndexedDB.open(name, description)
openSuccess():
db = event.target.result
request = db.setVersion('version 1')
cleanDatabase():
Deleted all object stores.
objectStore = db.createObjectStore('baz', { keyPath: 'id' });
Expecting exception from objectStore.add({id: 5}, 5);
PASS Exception was thrown.
PASS code is webkitIDBDatabaseException.DATA_ERR
PASS successfullyParsed is true

TEST COMPLETE
